# swagger.yml
{ 
   "swagger":"2.0",
   "info":{ 
      "description":"Apis for Parivahan Application API",
      "version":"1.0.0",
      "title":"Parivahan Application API",
      "contact":{ 
         "email":""
      },
      "license":{ 
         "name":"Apache 2.0",
         "url":"http://www.apache.org/licenses/LICENSE-2.0.html"
      }
   },
   "host":"localhost",
   "basePath":"/",
...
   "paths":{ 
...
      "/api/v1/ride":{ 
         "get":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"getAllRides",
            "description":"get All rides",
            "operationId":"getAllRides",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"Content-Type",
                  "in":"header",
                  "description":"Content-Type",
                  "required":true,
                  "type":"string",
                  "default":"application/json"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "type":"array",
                     "items":{ 
                        "$ref":"#/definitions/Ride"
                     }
                  }
               },
               "400":{ 
                  "description":"Bad Request."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error."
               }
            }
         },
         "post":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"offer ride",
            "description":"Host can offer ride using this api with his vehicle details and Start and end Location",
            "operationId":"offerRide",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"email",
                  "in":"query",
                  "description":"email",
                  "required":true,
                  "type":"string"
               },
               { 
                  "name":"Content-Type",
                  "in":"header",
                  "description":"Content-Type",
                  "required":true,
                  "type":"string",
                  "default":"application/json"
               },
               { 
                  "in":"body",
                  "name":"rideParams",
                  "description":"Ride Details",
                  "required":true,
                  "schema":{ 
                     "$ref":"#/definitions/Ride"
                  }
               },
               { 
                  "name":"showDetailsOnly",
                  "in":"query",
                  "description":"Show ride details only",
                  "required":false,
                  "type":"boolean",
                  "default":false
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK, Ride Added Successfully.",
                  "schema":{ 
                     "$ref":"#/definitions/Ride"
                  }
               },
               "201":{ 
                  "description":"Created"
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error"
               }
            }
         },
         "delete":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"Delete offered ride",
            "description":"Host can delete the offered ride using this api",
            "operationId":"DeleteRide",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"offerCode",
                  "in":"query",
                  "description":"offerCode",
                  "required":false,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK, Ride Added Successfully.",
                  "schema":{ 
                     "$ref":"#/definitions/Ride"
                  }
               },
               "204":{ 
                  "description":"No Content"
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error"
               }
            }
         }
      },
...
      "/api/v1/ride/find":{ 
         "get":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"get joinRides",
            "description":"Api will Return Matching rides as per current location and end location",
            "operationId":"joinRides",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"Content-Type",
                  "in":"header",
                  "description":"Content-Type",
                  "required":true,
                  "type":"string",
                  "default":"application/json"
               },
               { 
                  "name":"startLocation",
                  "in":"query",
                  "description":"Start location of Ride",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"endLocation",
                  "in":"query",
                  "description":"End Location of Ride",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"currentGPSLocation",
                  "in":"query",
                  "description":"Current GPS Location of rider",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"currency",
                  "in":"query",
                  "description":"Currnecy of country",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"startCoordinate",
                  "in":"query",
                  "description":"Start co-ordinate",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"endrtCoordinate",
                  "in":"query",
                  "description":"End co-ordinate",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"date",
                  "in":"query",
                  "description":"Date of Start journey",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"Hours",
                  "in":"query",
                  "description":"Time to Start journey",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"page",
                  "in":"query",
                  "description":"Page number to search",
                  "required":false,
                  "type":"integer",
                  "default":1,
                  "format":"int32"
               },
               { 
                  "name":"seats",
                  "in":"query",
                  "description":"Number of seats to search for ride",
                  "required":false,
                  "type":"integer",
                  "default":1,
                  "format":"int32"
               },
               { 
                  "name":"photo",
                  "in":"query",
                  "description":"Does photo of driver is filter parameter",
                  "required":false,
                  "type":"boolean",
                  "default":false
               },
               { 
                  "name":"sort",
                  "in":"query",
                  "description":"Sort rides by Date, nearest date first",
                  "required":false,
                  "type":"boolean",
                  "default":true
               },
               { 
                  "name":"order",
                  "in":"query",
                  "description":"Order rides as per charges, Lowest price first",
                  "required":false,
                  "type":"boolean",
                  "default":true
               },
               { 
                  "name":"radius",
                  "in":"query",
                  "description":"Radius from current location to search rides in Meters",
                  "required":false,
                  "type":"integer",
                  "default":0,
                  "format":"int32"
               },
               { 
                  "name":"radiusFromStartLocation",
                  "in":"query",
                  "description":"Radius from Start Location to search rides in Meters",
                  "required":false,
                  "type":"integer",
                  "default":0,
                  "format":"int32"
               },
               { 
                  "name":"radiusFromEndLocation",
                  "in":"query",
                  "description":"Radius from End Location to search rides in Meters",
                  "required":false,
                  "type":"integer",
                  "default":5,
                  "format":"int32"
               },
               { 
                  "name":"minPrice",
                  "in":"query",
                  "description":"Minimum price from which search will start",
                  "required":false,
                  "type":"integer",
                  "default":0,
                  "format":"int32"
               },
               { 
                  "name":"maxPrice",
                  "in":"query",
                  "description":"Maximum pricee for searching rides",
                  "required":false,
                  "type":"integer",
                  "default":5000,
                  "format":"int32"
               },
               { 
                  "name":"email",
                  "in":"query",
                  "description":"Email id of user",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "type":"array",
                     "items":{ 
                        "$ref":"#/definitions/Ride"
                     }
                  }
               },
               "400":{ 
                  "description":"Bad Request"
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error"
               }
            }
         }
      },
...
      "/api/v1/ride/{rideid}":{ 
         "post":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"get Ride Details",
            "description":"Api will Return ride details as per riders current Location",
            "operationId":"getRideDetails",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"rideid",
                  "in":"path",
                  "description":"rideid",
                  "required":true,
                  "type":"string"
               },
               { 
                  "in":"body",
                  "name":"rideParams",
                  "description":"ride details",
                  "required":true,
                  "schema":{ 
                     "$ref":"#/definitions/RideParams"
                  }
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "$ref":"#/definitions/RideSearch"
                  }
               },
               "201":{ 
                  "description":"Created"
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error."
               }
            }
         }
      },
      "/api/v1/shuttle/find/{date}":{ 
         "get":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"get shuttle Details",
            "description":"Api will Return shuttle details",
            "operationId":"getShuttleDetails",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"Content-Type",
                  "in":"header",
                  "description":"Content-Type",
                  "required":true,
                  "type":"string",
                  "default":"application/json"
               },
               { 
                  "name":"date",
                  "in":"path",
                  "description":"date",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "$ref":"#/definitions/RideSearch"
                  }
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Sevrer, Unknown Error."
               }
            }
         }
      },
...
      "/api/v1/user/booking":{ 
         "get":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"get shuttle Details",
            "description":"Api will Return shuttle details",
            "operationId":"getShuttleDetails_1",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"type",
                  "in":"query",
                  "description":"type",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"email",
                  "in":"query",
                  "description":"email",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "$ref":"#/definitions/RideSearch"
                  }
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Server, Unknown Error."
               }
            }
         }
      },
      "/api/v1/user/offer":{ 
         "get":{ 
            "tags":[ 
               "developers"
            ],
            "summary":"get shuttle Details",
            "description":"Api will Return shuttle details",
            "operationId":"getShuttleDetails_2",
            "consumes":[ 
               "application/json"
            ],
            "produces":[ 
               "application/json"
            ],
            "parameters":[ 
               { 
                  "name":"Authorization",
                  "in":"header",
                  "required":false,
                  "type":"string"
               },
               { 
                  "name":"email",
                  "in":"query",
                  "description":"email",
                  "required":false,
                  "type":"string"
               }
            ],
            "responses":{ 
               "200":{ 
                  "description":"OK.",
                  "schema":{ 
                     "$ref":"#/definitions/RideSearch"
                  }
               },
               "400":{ 
                  "description":"Bad Request, All parameters are mandatory."
               },
               "401":{ 
                  "description":"Unauthorized"
               },
               "403":{ 
                  "description":"Forbidden"
               },
               "404":{ 
                  "description":"Not Found."
               },
               "500":{ 
                  "description":"Intenal Server, Unknown Error."
               }
            }
         }
      },
...
   },
   "definitions":{ 
...
      "Host":{ 
         "type":"object",
         "properties":{ 
            "Rating":{ 
               "type":"string",
               "example":"4.3"
            },
            "hostAge":{ 
               "type":"integer",
               "format":"int32",
               "example":26
            },
            "hostGender":{ 
               "type":"string",
               "example":"Male"
            },
            "hostName":{ 
               "type":"string",
               "example":"Swapnil"
            },
            "isGovtIdverified":{ 
               "type":"boolean",
               "example":false
            },
            "isPhoneNumberVerified":{ 
               "type":"boolean",
               "example":true
            },
            "totalHostRatings":{ 
               "type":"integer",
               "format":"int32",
               "example":6
            }
         }
      },
      "LatLng":{ 
         "type":"object",
         "properties":{ 
            "lat":{ 
               "type":"number",
               "format":"double"
            },
            "lng":{ 
               "type":"number",
               "format":"double"
            }
         }
      },
...
      "Ride":{ 
         "type":"object",
         "required":[ 
            "availableSeats",
            "charges",
            "departureDate",
            "departureHours",
            "endCoordinate",
            "startCoordinate"
         ],
         "properties":{ 
            "availableSeats":{ 
               "type":"integer",
               "format":"int32",
               "example":3
            },
            "charges":{ 
               "type":"integer",
               "format":"int32",
               "example":100
            },
            "departureDate":{ 
               "type":"string",
               "example":"2019-02-2T07:46:22:427Z"
            },
            "departureHours":{ 
               "type":"string",
               "example":"0000"
            },
            "encodedPath":{ 
               "type":"string",
               "example":"bgcduh435ry78hjshc984yfhichkidf4yiuwbpxjdskcn"
            },
            "endCoordinate":{ 
               "type":"string",
               "example":"19.2183,72.9781"
            },
            "endLatLng":{ 
               "$ref":"#/definitions/LatLng"
            },
            "endLocation":{ 
               "type":"string",
               "example":"Pune University"
            },
            "host":{ 
               "$ref":"#/definitions/Host"
            },
            "id":{ 
               "type":"string",
               "example":"1200"
            },
            "ridePublishedDate":{ 
               "type":"string",
               "example":"15-08-2018"
            },
            "ridePublishedTime":{ 
               "type":"string",
               "example":"1600"
            },
            "rideRequests":{ 
               "type":"array",
               "items":{ 
                  "$ref":"#/definitions/RideRequestVO"
               }
            },
            "selectedRidePath":{ 
               "type":"string"
            },
            "startCoordinate":{ 
               "type":"string",
               "example":"18.5204,73.8567"
            },
            "startLatLng":{ 
               "$ref":"#/definitions/LatLng"
            },
            "startLocation":{ 
               "type":"string",
               "example":"Gslab pune"
            },
            "stopOvers":{ 
               "type":"array",
               "items":{ 
                  "type":"string"
               }
            },
            "totalRidesOffered":{ 
               "type":"integer",
               "format":"int32",
               "example":18
            },
            "totalSeats":{ 
               "type":"integer",
               "format":"int32",
               "example":4
            },
            "user":{ 
               "$ref":"#/definitions/UserVO"
            },
            "vehicle":{ 
               "type":"string",
               "enum":[ 
                  "car",
                  "bike",
                  "shuttle"
               ]
            }
         }
      },
      "RideParams":{ 
         "type":"object",
         "properties":{ 
            "destinationWalkingDistance":{ 
               "type":"string",
               "example":"1.5km"
            },
            "destinationWalkingDistanceInMeters":{ 
               "type":"integer",
               "format":"int64",
               "example":1500
            },
            "originWalkingDistance":{ 
               "type":"string",
               "example":"1.5km"
            },
            "originWalkingDistanceInMeters":{ 
               "type":"integer",
               "format":"int64",
               "example":1500
            },
            "riderDestinationCoordinate":{ 
               "type":"string",
               "example":"18.554230,73.809636"
            },
            "riderJoiningPointCoordinate":{ 
               "type":"string",
               "example":"18.550069,73.809797"
            },
            "totalChargesForRide":{ 
               "type":"integer",
               "format":"int32",
               "example":50
            },
            "totalRideDistance":{ 
               "type":"string",
               "example":"5000"
            },
            "totalTravelDurationInSeconds":{ 
               "type":"string",
               "example":"2000"
            }
         }
      },
...
      "RideRequestVO":{ 
         "type":"object",
         "properties":{ 
            "bookingId":{ 
               "type":"string"
            },
            "endLocation":{ 
               "type":"string"
            },
            "ride":{ 
               "$ref":"#/definitions/Ride"
            },
            "rideCharges":{ 
               "type":"integer",
               "format":"int32"
            },
            "rideId":{ 
               "type":"string"
            },
            "startLocation":{ 
               "type":"string"
            },
            "status":{ 
               "type":"string",
               "enum":[ 
                  "ACCEPTED",
                  "DECLINED",
                  "PENDING",
                  "CANCELLED"
               ]
            },
            "user":{ 
               "$ref":"#/definitions/UserVO"
            },
            "userEmail":{ 
               "type":"string"
            }
         }
      },
      "RideSearch":{ 
         "type":"object",
         "properties":{ 
            "destinationWalkingDistance":{ 
               "type":"string",
               "example":"1.5km"
            },
            "destinationWalkingDistanceInMeters":{ 
               "type":"integer",
               "format":"int64",
               "example":1500
            },
            "originWalkingDistance":{ 
               "type":"string",
               "example":"1.5km"
            },
            "originWalkingDistanceInMeters":{ 
               "type":"integer",
               "format":"int64",
               "example":1500
            },
            "ride":{ 
               "$ref":"#/definitions/Ride"
            },
            "riderEndPointCoordinate":{ 
               "$ref":"#/definitions/LatLng"
            },
            "riderJoiningPointCoordinate":{ 
               "example":"18.550069,73.809797",
               "$ref":"#/definitions/LatLng"
            },
            "riderOriginCoordinate":{ 
               "type":"string",
               "example":"18.554230,73.809636"
            },
            "totalChargesForRide":{ 
               "type":"integer",
               "format":"int32",
               "example":50
            },
            "totalRideDistance":{ 
               "type":"string",
               "example":"5000"
            },
            "totalTravelDurationInSeconds":{ 
               "type":"string",
               "example":"2000"
            },
            "walkingDurationInSeconds":{ 
               "type":"string",
               "example":"720"
            },
            "walkingEncodedPath":{ 
               "type":"string",
               "example":"fywefg469ncjd@3mld@xekjd2w3$5%vfsd"
            }
         }
      },
...
   }
}